feat(product): add completed product service (#41)
* Im going to sleep

* fix(product): missing postgres driver

* refactor(product): change int64 to int32 in proto files for api gateway

* fix(product): add missing database schema creation

* fix(product): move methods in same directory

This allows the server to now see the different methods available

---------

+++ productService/methods_services/product/get.go
@@ -14,7 +14,7 @@ func GetProductByID(conf config.Config, id int) (*product.Product, error) {
 	}
 
 	return &product.Product{
-		Id:          int64(p.ID),
+		Id:          int32(p.ID),
 		Name:        p.Name,
 		Description: p.Description,
 		Price:       p.Price,
